% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmultinom.R
\name{handle_overflow}
\alias{handle_overflow}
\title{Return the correct p-value in the case of an overflow error}
\usage{
handle_overflow(lower.tail)
}
\description{
Return the correct p-value in the case of an overflow error
}
