% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmultinom.R
\name{calculate_upper_tail}
\alias{calculate_upper_tail}
\title{Convert the p-value for the lower tail to the upper tail p-value}
\usage{
calculate_upper_tail(plower)
}
\arguments{
\item{plower}{float; The lower tail p-value}
}
\value{
float; The upper tail p-value
}
\description{
Convert the p-value for the lower tail to the upper tail p-value
}
